Rename truffle-code-utils to @truffle/code-utils
