+++ packages/code-utils/package.json
@@ -1,12 +1,12 @@
 {
-  "name": "truffle-code-utils",
+  "name": "@truffle/code-utils",
   "version": "1.2.4",
   "description": "Utilities for parsing and managing EVM-compatible bytecode",
   "main": "index.js",
   "scripts": {
     "test": "mocha"
   },
-  "repository": "https://github.com/trufflesuite/truffle/tree/master/packages/truffle-code-utils",
+  "repository": "https://github.com/trufflesuite/truffle/tree/develop/packages/code-utils",
   "author": "",
   "license": "MIT",
   "publishConfig": {
