+++ packages/truffle-core/lib/commands/opcode.js
@@ -21,7 +21,7 @@ var command = {
     var Config = require("truffle-config");
     var TruffleError = require("@truffle/error");
     var Contracts = require("truffle-workflow-compile");
-    var CodeUtils = require("truffle-code-utils");
+    var CodeUtils = require("@truffle/code-utils");
 
     if (options._.length === 0) {
       return done(new TruffleError("Please specify a contract name."));
