+++ packages/truffle-debugger/package.json
@@ -30,7 +30,7 @@
     "redux-saga": "1.0.0",
     "remote-redux-devtools": "^0.5.12",
     "reselect-tree": "^1.3.1",
-    "truffle-code-utils": "^1.2.4",
+    "@truffle/code-utils": "^1.2.4",
     "truffle-decode-utils": "^1.0.16",
     "truffle-decoder": "^3.0.10",
     "truffle-expect": "^0.0.9",
